backend/pyproject.toml
[project]
name = "mask-generator"
version = "0.1.0"
description = "A tool for generating masks from images"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "pillow>=10.0.0",
    "django-cors-headers>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mask_generator.settings"
python_files = "test_*.py"
testpaths = ["api/tests"]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88